Add Salesforce Apex support
This uses aheber/tree-sitter-sfapex as it is the most complete
implementation of a tree-sitter grammar for Apex, SOQL and SOSL.
It has been tested on a large set of code and is used by other
major products (see https://github.com/aheber/tree-sitter-sfapex/issues/4).

This repository contains 3 grammars: Apex, SOQL and SOSL. Apex is the
main coding language. SOQL and SOSL are 2 database query languages which
can be used independentely or embedded in Apex code.